Move OidcSessionRegistry Login Configuration to oauth2Login
